pyproject.toml
[build-system]
requires = [
    "setuptools",
    "setuptools_scm[toml]",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[project]
name = "RegisterImages"
dynamic = ["version"]
readme = "README.rst"
classifiers = ["License :: OSI Approved :: Apache Software License"]
requires-python = ">=3.8"
dependencies = [
    "girder-client",
    "histomicstk",
    "large-image[sources]",
    "large-image-converter",
    "scikit-image",
    "scikit-learn",
    "pyarrow",
    "pystackreg",
    "rasterio",
    "shapely",
    "umap-learn",
]
